refactor: repackage redis configure and service
