+++ module-api/src/main/java/com/codingbottle/common/redis/config/RedisConfig.java
@@ -1,4 +1,4 @@
-package com.codingbottle.common.config;
+package com.codingbottle.common.redis.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
