+++ module-api/src/main/java/com/codingbottle/domain/weather/service/WeatherService.java
@@ -3,7 +3,7 @@ package com.codingbottle.domain.weather.service;
 import com.codingbottle.auth.entity.User;
 import com.codingbottle.common.exception.ApplicationErrorException;
 import com.codingbottle.common.exception.ApplicationErrorType;
-import com.codingbottle.common.redis.WeatherRedisService;
+import com.codingbottle.common.redis.service.WeatherRedisService;
 import com.codingbottle.domain.region.entity.Region;
 import com.codingbottle.domain.weather.model.CurrentWeatherRequest;
 import com.codingbottle.domain.weather.model.WeatherResponse;
